+++ example/lib/home_view.dart
@@ -7,6 +7,7 @@ import 'package:example/game_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
+import 'package:squares/squares.dart';
 
 class HomeView extends StatefulWidget {
   HomeView({Key? key}) : super(key: key);
@@ -73,6 +74,7 @@ class _HomeViewState extends State<HomeView> with WidgetsBindingObserver {
                             game: games[i],
                             theme: state.theme,
                             pieceSet: state.pieceSet,
+                            highlightTheme: state.highlightTheme,
                           );
                           return Card(
                             child: Stack(